Add model.go that has the struct for the new notepad
